Refactor gateway package structure
