+++ base/gateway/src/runtime/ros_topic_relay.js
@@ -2,7 +2,7 @@
 
 const { encodeBaseRtcm, encodeBaseSvin } = require('../protocol/sik')
 
-async function startRosBridge(options = {}) {
+async function startRosTopicRelay(options = {}) {
   const nextSeq = options.nextSeq
   const writeFrame = options.writeFrame
   const log = typeof options.log === 'function' ? options.log : () => {}
@@ -15,7 +15,7 @@ async function startRosBridge(options = {}) {
       // eslint-disable-next-line global-require
       rclnodejs = require('rclnodejs')
     } catch (err) {
-      log(`ROS bridge disabled: rclnodejs not available (${err.message})`)
+      log(`ROS topic relay disabled: rclnodejs not available (${err.message})`)
       return { stop() {} }
     }
   }
@@ -28,13 +28,13 @@ async function startRosBridge(options = {}) {
     }
   } catch (err) {
     if (!/already been initialized/i.test(err.message || '')) {
-      log(`ROS bridge init failed: ${err.message}`)
+      log(`ROS topic relay init failed: ${err.message}`)
       return { stop() {} }
     }
   }
 
   let lastSvinTxMs = 0
-  const nodeName = `gateway_ros_${process.pid || Math.floor(Math.random() * 1e5)}`
+  const nodeName = `gateway_ros_topics_${process.pid || Math.floor(Math.random() * 1e5)}`
   const rosNode = new rclnodejs.Node(nodeName)
 
   rosNode.createSubscription(
@@ -76,7 +76,7 @@ async function startRosBridge(options = {}) {
   })
 
   rclnodejs.spin(rosNode)
-  log('ROS bridge started (SVIN + RTCM over SiK)')
+  log('ROS topic relay started (SVIN + RTCM over SiK)')
 
   return {
     async stop() {
@@ -102,5 +102,5 @@ async function startRosBridge(options = {}) {
 }
 
 module.exports = {
-  startRosBridge,
+  startRosTopicRelay,
 }
